rename exceptions to _exceptions
