+++ qupid/_exceptions.py
